Add `:slug` modifier for template variables
This modifier will sanitize the variable's value, replacing all
sequences of non-word characters (other than _-) with a
single - character. This allows, for example, certain characters
(like periods) to be allowed in the filename, but not as part of
a particular variable substitution. Also lowercases the value.

Originally introduced for Dendron compatibility.