support conditional substitions in templates
This adds a simple syntax (see options.html for documentation)
to conditionally include a string in the template IF (or UNLESS)
a certain template value is present. This is an extension of the
existing template modifier syntax.

Can be used to optionally insert delimiters if a value is present,
or to provide replacement ("default") values if it's absent.

Originally added for Dendron compatibility.