add {id} variable for per-capture unique IDs
When capturing a page, generates a UUID and exposes it via the
`{id}` variable. Two different captures of the same page will have
different IDs, but all references to `{id}` will return _the same_ ID
within the context of an individual capture.

The ID is intended to be an opaque alphanumeric string. It's
generated by calling randomUUID() and removing the dashes.

Originally added for Dendron compatibility.