don't mutate the actual document
I encountered an issue where removeHiddenNodes() actually
removes nodes not just from the HTML used to render the MD,
but actually changing the tab content itself. On some sites with
SVG icons with aria-hidden=true, it's very obvious because the
icons all disappear. However, this seems to only happen when
side-loading the extension with about:debugging. I don't really
know enough about extension development to understand why
this issue would only appear in dev. The fix I implemented is to
clone the document into a fragment, and mutate the fragment.